Refactor: Jacoco report aggregation 플러그인 적용
